% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.IsingFitBO.R
\name{print.IsingFitBO}
\alias{print.IsingFitBO}
\title{Print Method for IsingFitBO Objects}
\usage{
\method{print}{IsingFitBO}(x, ...)
}
\arguments{
\item{x}{An object of class `IsingFitBO`}

\item{...}{Additional arguments}
}
\value{
Invisibly returns the input object.
}
\description{
Displays concise output of Ising model estimation results.
}
